- refactoring the alexa skill
