+++ src/typescript/handler/CallbackHandler.ts
@@ -1,0 +1,9 @@
+import {SpeechResponseHandler} from "./SpeechResponseHandler";
+
+export class CallbackHandler {
+    getAlexaCallback(sessionAttributes, speechletResponse, callback) {
+        let speechResponseHandler = new SpeechResponseHandler();
+        callback(null, speechResponseHandler.buildResponse(sessionAttributes, speechletResponse));
+        return callback;
+    }
+}
