+++ src/typescript/type/AlexaEventType.ts
@@ -1,0 +1,6 @@
+export enum AlexaEventType {
+    LAUNCH_REQUEST = 'LaunchRequest',
+    INTENT_REQUEST = 'IntentRequest',
+    SESSION_ENDED = 'SessionEndedRequest'
+}
+
